requirements.txt
discord.py==1.4.1
python-decouple==3.3
shortuuid==1.0.1
pychal==1.10.0